fix(client/server): Stored vehicle issue if using int structure MYSQL
https://github.com/renzuzu/renzu_garage/issues/233
https://github.com/renzuzu/renzu_garage/issues/235
fixes( real park issue if model is missing in image )
https://github.com/renzuzu/renzu_garage/issues/232

+++ garage.sql
@@ -5,7 +5,7 @@ ALTER TABLE owned_vehicles
 ADD impound int(1) NOT NULL DEFAULT 0;
 
 ALTER TABLE owned_vehicles
-ADD `stored` int(1) NOT NULL DEFAULT 0;
+ADD `stored` tinyint(1) NOT NULL DEFAULT 0;
 
 ALTER TABLE owned_vehicles
 ADD `type` varchar(32) NOT NULL DEFAULT 'car';
